Introduce TestUtils and refactor test
